// src/sanity/types.ts
export type Notes = {
  _id: string;
  _type: "notes";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  author?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "author";
  };
  views?: number;
  country?: string;
  universityName?: string;
  description?: string;
  category?: string;
  image?: string;
  file?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.fileAsset";
    };
    media?: unknown;
    _type: "file";
  };
};

export type Author = {
  _id: string;
  _type: "author";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  id?: number;
  name?: string;
  username?: string;
  email?: string;
  image?: string;
  bio?: string;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./src/lib/queries.ts
// Variable: NOTES_QUERY
// Query: *[  _type == "notes" && defined(slug.current)] | order(_createdAt desc) {  _id,   _type,  _createdAt,  _updatedAt,  _rev,  title,   slug,  _createdAt,  author -> {    _id,    name,    image,    bio  },   views,  description,  category,  country,  universityName,  image,  file,}
export type NOTES_QUERYResult = Array<{
  _id: string;
  _type: "notes";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title: string | null;
  slug: Slug | null;
  author: {
    _id: string;
    name: string | null;
    image: string | null;
    bio: string | null;
    country?: string | null;
  universityName?: string | null;
  } | null;
  views: number | null;
  description: string | null;
  category: string | null;
  country: string | null;
  universityName: string | null;
  image: string | null;
  file: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.fileAsset";
    };
    media?: unknown;
    _type: "file";
  } | null;
}>;